Extract Diary to a separate file
